MS-65: implement React Lazy
