+++ src/Components/Admin/Admins/index.js
@@ -1,61 +1,64 @@
 import { useEffect, useState } from 'react';
-import styles from './councelors.module.css';
+import styles from './admins.module.css';
 import Modal from 'Components/Shared/Modal';
 import ErrorModal from 'Components/Shared/ErrorModal';
 import IsLoading from 'Components/Shared/IsLoading/IsLoading';
 import Button from 'Components/Shared/Button/Button';
 import DeleteButton from 'Components/Shared/DeleteButton/DeleteButton';
 import { useSelector, useDispatch } from 'react-redux';
-import { getCounselors, deleteCounselor } from 'redux/counselors/thunks';
+import { getAdmins, deleteAdmin } from 'redux/admins/thunks';
 import { useHistory } from 'react-router-dom';
-import { errorToDefault } from 'redux/counselors/actions';
+import { errorToDefault } from 'redux/admins/actions';
 
-const Councelor = () => {
+const Admins = () => {
   const [showModal, setShowModal] = useState(false);
   const [selectedId, setSelectedId] = useState('');
+
   const history = useHistory();
 
   const dispatch = useDispatch();
 
-  const counselors = useSelector((store) => store.counselors.list);
-  const isLoading = useSelector((store) => store.counselors.isLoading);
-  const error = useSelector((store) => store.counselors.error);
-  const errorMessage = useSelector((store) => store.counselors.errorMessage);
+  const admins = useSelector((store) => store.admins.list);
+  const isLoading = useSelector((store) => store.admins.isLoading);
+  const error = useSelector((store) => store.admins.error);
+  const errorMessage = useSelector((store) => store.admins.errorMessage);
 
   useEffect(() => {
-    if (!counselors.length) {
-      dispatch(getCounselors());
+    if (!admins.length) {
+      dispatch(getAdmins());
     }
-  }, [counselors]);
+  }, [admins]);
 
   const onClickDelete = () => {
-    dispatch(deleteCounselor(selectedId));
+    dispatch(deleteAdmin(selectedId));
     setShowModal(false);
-  };
-
-  const handleIdCouncelor = (event, id) => {
-    event.stopPropagation();
-    setShowModal(true);
-    setSelectedId(id);
   };
 
   const closeModal = () => {
     setShowModal(false);
   };
 
-  if (isLoading) return <IsLoading />;
+  const handleIdAdmin = (event, id) => {
+    event.stopPropagation();
+    setSelectedId(id);
+    setShowModal(true);
+  };
+
+  if (isLoading) {
+    return <IsLoading />;
+  }
 
   return (
-    <section className={styles.container}>
+    <div className={styles.container}>
       <Modal
         showModal={showModal}
         closeModal={closeModal}
         actionEntity={onClickDelete}
         selectedId={selectedId}
-        titleText="Delete a counselor"
+        titleText="Delete an Admin"
         spanObjectArray={[
           {
-            span: 'Are you sure you want to delete this counselor?'
+            span: 'Are you sure you want to delete this admin?'
           }
         ]}
         leftButtonText="delete"
@@ -69,37 +72,36 @@ const Councelor = () => {
         buttonText="ok"
       />
       <div className={styles.titleAndButton}>
-        <h3>Councelors</h3>
-        <Button onClick={() => history.push('/councelors/form')} value="Counselor" />
+        <h3>Admin</h3>
+        <Button onClick={() => history.push('/admin/admins/form')} value="Admin" />
       </div>
       <table className={styles.list}>
         <thead>
           <tr>
-            <th> First Name </th>
-            <th> Last Name </th>
+            <th> Email </th>
             <th> Actions </th>
           </tr>
         </thead>
         <tbody>
-          {counselors.map((counselor) => {
+          {admins.map((admin) => {
             return (
               <tr
-                key={counselor._id}
+                key={admin._id}
                 onClick={() => {
-                  window.location.replace(`councelors/form?id=${counselor._id}`);
+                  window.location.replace(`admins/form?id=${admin._id}`);
                 }}
               >
-                <td>{counselor.firstName}</td>
-                <td>{counselor.lastName}</td>
+                <td>{admin.email}</td>
                 <td>
-                  <DeleteButton onClick={(event) => handleIdCouncelor(event, counselor._id)} />
+                  <DeleteButton onClick={(event) => handleIdAdmin(event, admin._id)} />
                 </td>
               </tr>
             );
           })}
         </tbody>
       </table>
-    </section>
+    </div>
   );
 };
-export default Councelor;
+
+export default Admins;
